Add DigitalOcean target for PMM2 (#296)
* Add DigitalOcean target for PMM2

Previously [opened in percona-images](https://github.com/Percona-Lab/percona-images/pull/89).

This adds a Packer target for building a DigitalOcean image for the purpose of listing in DigitalOcean's Marketplace.

Due to the lack of a DigitalOcean-equivalent to the amazon-ebs Packer builder, it was necessary to build the DigitalOcean image on. single volume (instance storage). To this end, a single_disk variable was added to the lvm-init role to gate tasks that don't apply to a single-volume install.

A helper script, digitalocean_add_dbaas, is included in the admin user's home directory on DigitalOcean images to assist in adding DigitalOcean DBaaS MySQL instances to PMM.

* Use larger VM type

* Use `block` to eliminate repeated conditions

* Add `pmm_client_repo_name` declaration